Refactor configuration and update index performance handling
- Added application.yml for test profile and datasource configuration.
- Updated AutoSyncScheduler to use dynamic cron expression.
- Modified DashboardService and related classes to support BigDecimal for baseIndex.
- Enhanced performance ranking logic to include optional indexInfoId filter.

+++ src/main/java/com/findex/dto/indexinfo/IndexInfoDto.java
@@ -1,6 +1,7 @@
 package com.findex.dto.indexinfo;
 
 import com.findex.enums.IndexSourceType;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record IndexInfoDto(
@@ -9,7 +10,7 @@ public record IndexInfoDto(
     String indexName,
     Integer employedItemsCount,
     LocalDate basePointInTime,
-    Integer baseIndex,
+    BigDecimal baseIndex,
     IndexSourceType sourceType,
     boolean favorite
 ) {}
