+++ src/main/java/com/findex/dto/indexinfo/IndexInfoUpdateRequest.java
@@ -1,10 +1,11 @@
 package com.findex.dto.indexinfo;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record IndexInfoUpdateRequest(
     Integer employedItemsCount,
     LocalDate basePointInTime,
-    Integer baseIndex,
+    BigDecimal baseIndex,
     Boolean favorite
 ) {}
