+++ src/main/java/com/findex/dto/syncjob/OpenApiIndexInfoItem.java
@@ -1,9 +1,11 @@
 package com.findex.dto.syncjob;
 
+import java.math.BigDecimal;
+
 public record OpenApiIndexInfoItem(
     String idxCsf,
     String idxNm,
     Integer epyItmsCnt,
     String basPntm,
-    Integer basIdx
+    BigDecimal basIdx
 ) {}
